cqf/requirements.txt
numpy>=1.21.0
pytorch>=1.10.0
matplotlib>=3.4.0
optuna>=3.0.0
joblib>=1.1.0
plotly>=5.0.0
scipy>=1.7.0